Remove references to receipts
I’d advise against mentioning receipts/invoices. Sponsors don’t want to
get into the business of handling his kind of paperwork or approving
something with a specific limit.

Set the rate at something you can live with and be prepared to “float”
a loan to the organization and be under budget for the next meeting.