Complete re-write of the sponsorship page
I want to say that we are looking for “partners.”

I wanted to call out advantage for both sponsorship and providing
speakers to make it clear that we are looking for supporting
relationship and to make it seem less mercenary (we aren’t just looking
for money).

I wanted to be more focused in the favorables to the company and the
employee.

I want to emphasize the recruitment aspect (by putting it first), but
not talk about specific amounts or percentages.

I wanted to call out exactly what companies get for being sponsors.