Rewrite of sponsorship page
Wanted to emphasize that we are looking for partners.

Wanted to ask for both presenters and sponsorship to make it seem more
like a partnership and less mercenary (we aren’t just asking for money).

Wanted to emphasize recruitment (by putting it first), but not mention
any specific amounts or proportions.

Wanted to call out specific favorables for both companies and
individual attendees (and make them focused).

Wanted to list specific benefits of the sponsorship.